split lightrag_visualizer into separate module and update entry point
